app/requirements.txt
cachetools==3.1.0
Click==7.0
Flask==1.0.2
google-api-python-client==1.7.8
google-auth==1.6.3
google-auth-httplib2==0.0.3
httplib2==0.12.3
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
numpy==1.16.3
oauth2client==4.1.3
pandas==0.24.2
pyasn1==0.4.5
pyasn1-modules==0.2.5
pyodbc==4.0.26
python-dateutil==2.8.0
pytz==2019.1
rsa==4.0
six==1.12.0
uritemplate==3.0.0
Werkzeug==0.15.2
simplejson==3.17.0
mysql-connector-python==8.0.5
pymysql==1.0.2
